Implement god mode with a cookie as the storage
The concept of god mode is simple: override the user with some
other user so the app can present itself as it was viewed with the
eyes of someone else.

The tricky bit is where do you store this state.
My original idea was to simply pass `godMode` as a request
paramter so one-off switch to another user is easy to perform.
I realized this wouldn't work because single view (e.g. of index)
spuns out more requests (e.g. with swr) and you need to thread
the state through. Therefore, I decided to reach out for more
persistent state storage: cookies.

We have an api endpoint that manipulates the state of godMode:

http://localhost:3002/api/godMode?as_user=a.magalska@nencki.edu.pl
to turn it on

http://localhost:3002/api/godMode?as_user=off
to turn it off

Later on, we could implement a little UI to call that API endpoint.